Separate gui into several files and cleanup code bases.
